Signup and Login working
Created the login.php script to handle login. Replaced depricated ProgressDialog with ProgressBar in the LoginActivity.kt class (yet to do this in the SignupActivity.java class). Also changed the JSONParser.java class to replace the depricated DefaultHttpClient object with the HttpURLConnection object (class is copied from: http://danielnugent.blogspot.se/ ; see the link for futher description). The Signup and Login function is now working but some alteration need to be made (e.g. caching so that one does not always need to login and adding social media login option).