Firebase signup (almost) ready
It is possible now signup by entering email and password. The user id and email will then be stored in the JSON database users in Firebase. There is a bug with signing in with google which needs to be fixed. Next steps is to implement the sign in function and retrieving the user data from the database.